setup react router, added navbar to root page
