+++ FrontEnd/datta-frontend/src/App.css
@@ -1,0 +1,4 @@
+.App{
+  /* to account for the navbar */
+  margin-left:90px;
+}
